requirements.txt
aiogram~=3.13.1
tortoise-orm
pydantic
aerich